requirements.txt
Flask
requests
python-Wappalyzer
beautifulsoup4
cryptography
setuptools